@@ -14,22 +14,31 @@
     }
   },
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
+    "psr/log": "^1.1||^2.0||^3.0",
     "twig/twig": "^2.0||^3.0",
-    "guzzlehttp/guzzle": "^6.0"
+    "guzzlehttp/guzzle": "^6.0||^7.0"
   },
   "require-dev": {
+    "doctrine/dbal": "~3.2",
+    "doctrine/migrations": "^3.3",
+    "elasticsearch/elasticsearch": "^6.0||^7.0",
+    "mongodb/mongodb": "^1.4",
+    "monolog/monolog": "^2.3",
+    "phpstan/phpstan": "^1.0",
     "phpunit/phpunit": "^9.0",
-    "symfony/browser-kit": "^5.0",
-    "symfony/css-selector": "^5.0",
-    "doctrine/dbal": "~2.2",
-    "elasticsearch/elasticsearch": "*",
-    "mongodb/mongodb": "^1.4"
+    "symfony/browser-kit": "^5.0||^6.0",
+    "symfony/css-selector": "^5.0||^6.0",
+    "symfony/event-dispatcher": "^5.0||^6.0"
   },
   "suggest": {
     "doctrine/dbal": "for checking DB connections",
+    "doctrine/migrations": "for checking Doctrine Migration status",
     "elasticsearch/elasticsearch": "for checking Elasticsearch",
     "mongodb/mongodb": "for checking MongoDB connections"
+  },
+  "config": {
+    "sort-packages": true
   },
   "scripts": {
     "test": "phpunit"
